requirements.txt
beautifulsoup4
requests
backoff
geopy
pycep-correios
python-Levenshtein
Unidecode
kafka-python
folium
mysql-connector-python
pandas
SQLAlchemy
pymysql
